requirements.txt
py-consul
docker
requests